Alert
* Alert to the delete

+++ confusion/components/FavoriteComponent.js
@@ -1,11 +1,16 @@
 import React, { Component } from 'react';
-import { View, FlatList, Text } from 'react-native';
+import { View, FlatList, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { baseUrl } from '../shared/baseUrl';
 import { Loading } from './LoadingComponent';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
+
+
+/*Suppressing the warning */
+console.disableYellowBox = true;
+/*Suppressing the warning */
 
 const mapStateToProps = state =>{
   return{
@@ -34,7 +39,25 @@ class Favorites extends Component{
           {
             text: 'Delete',
             type: 'delete',
-            onPress: () => this.props.deleteFavorite(item.id)
+            onPress: () => {
+              Alert.alert(
+                'Delete Favorite?',
+                'Are You sure you wish to delete the favorite dish ' + item.name + '?',
+                [
+                   {
+                   text: 'Cancel',
+                   onPress: () => console.log(item.name + 'Not Deleted'),
+                   style: '  cancel'
+                 },
+                  {
+                    text: 'OK',
+                    onPress: () => this.props.deleteFavorite(item.id)
+                  }
+                ],
+                { cancelable: false }
+              );
+            }
+
           }
         ];
